installing cache manager e redis to application

+++ prisma/migrations/20250616120921_init/migration.sql
@@ -1,6 +1,6 @@
 -- CreateTable
 CREATE TABLE "Admin" (
-    "id" BIGSERIAL NOT NULL,
+    "id" SMALLSERIAL NOT NULL,
     "created_at" TIMESTAMPTZ(6) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "username" VARCHAR NOT NULL DEFAULT '',
     "password" VARCHAR NOT NULL,
@@ -10,13 +10,13 @@ CREATE TABLE "Admin" (
 
 -- CreateTable
 CREATE TABLE "Users" (
-    "id" BIGSERIAL NOT NULL,
+    "id" SMALLSERIAL NOT NULL,
     "created_at" TIMESTAMPTZ(6) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "name" VARCHAR NOT NULL DEFAULT '',
     "email" VARCHAR NOT NULL DEFAULT '',
     "telephone" VARCHAR NOT NULL DEFAULT '',
     "role" VARCHAR NOT NULL DEFAULT '',
-    "age" BIGINT NOT NULL,
+    "age" SMALLINT NOT NULL,
     "salary" VARCHAR NOT NULL DEFAULT '',
 
     CONSTRAINT "Users_pkey" PRIMARY KEY ("id")
